`size_t` in `mojo::DataPipe[Consumer|Producer]Handle`: /device.
This CL makes use of the new `size_t`-based APIs in the remaining
directories like `//components`, `//extensions`, or `//storage`.
See https://crrev.com/c/5414148 for motivation.

This CL may have introduced new integer conversions - in those cases the
CL attempts to have well-defined behavior (e.g. using
`base::checked_cast`).  OTOH, this CL did *not* attempt to fix old
integer operations (e.g. old `int` => `uint32_t` / `size_t` conversions,
or old additions that may potentially overflow).

This CL was uploaded by git cl split.

R=mattreynolds@chromium.org

Bug: 40284755
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5462184
Auto-Submit: Łukasz Anforowicz <lukasza@chromium.org>
Commit-Queue: Łukasz Anforowicz <lukasza@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1288883}

+++ device/bluetooth/socket.cc
@@ -12,6 +12,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #include "base/functional/bind.h"
 #include "base/functional/callback_helpers.h"
 #include "base/memory/ptr_util.h"
+#include "base/numerics/safe_conversions.h"
 #include "device/bluetooth/bluetooth_socket.h"
 #include "mojo/public/cpp/bindings/pending_receiver.h"
 #include "mojo/public/cpp/bindings/receiver.h"
@@ -91,7 +92,7 @@ void Socket::ReceiveMore() {
 
   // Passing 0 as the initial value allows |pending_write_buffer_max_size| to be
   // assigned the buffer's max size.
-  uint32_t pending_write_buffer_max_size = 0;
+  size_t pending_write_buffer_max_size = 0;
 
   MojoResult result = receive_stream_->BeginWriteData(
       &pending_write_buffer, &pending_write_buffer_max_size,
@@ -105,7 +106,7 @@ void Socket::ReceiveMore() {
   }
 
   bluetooth_socket_->Receive(
-      pending_write_buffer_max_size,
+      base::checked_cast<int>(pending_write_buffer_max_size),
       base::BindOnce(&Socket::OnBluetoothSocketReceive,
                      weak_ptr_factory_.GetWeakPtr(), pending_write_buffer),
       base::BindOnce(&Socket::OnBluetoothSocketReceiveError,
@@ -165,7 +166,7 @@ void Socket::SendMore() {
   // Passing 0 as the initial value allows |pending_read_buffer_size| to be
   // assigned the number of bytes that the other side of |send_stream_| has
   // already written.
-  uint32_t pending_read_buffer_size = 0;
+  size_t pending_read_buffer_size = 0;
 
   MojoResult result = send_stream_->BeginReadData(&pending_read_buffer,
                                                   &pending_read_buffer_size,
@@ -181,8 +182,8 @@ void Socket::SendMore() {
   bluetooth_socket_->Send(
       base::MakeRefCounted<net::WrappedIOBuffer>(
           base::make_span(static_cast<const char*>(pending_read_buffer),
-                          static_cast<size_t>(pending_read_buffer_size))),
-      pending_read_buffer_size,
+                          pending_read_buffer_size)),
+      base::checked_cast<int>(pending_read_buffer_size),
       base::BindOnce(&Socket::OnBluetoothSocketSend,
                      weak_ptr_factory_.GetWeakPtr()),
       base::BindOnce(&Socket::OnBluetoothSocketSendError,
